Refactor code and prepare for new functionality
Three new pieces of quality of life functionality will be added to the
clock after using it for a few years. New features are as follows:
ability to change the amount of work sets before the rest timer goes,
ability to make any work set as the last, and a toggle to stop the next
work set from starting automatically after a break.